build: Handle program targets correctly
Due to the fact that platforms disagree about which file-name
extension an executable should have, subtle problems have accrued;
notably, it became the case that build targets were no longer
explicitly being treated as actual, real files, but sometimes as
implicitly phony targets (that is, a target which should have been
referring to a file actually wasn't, because it was expressed
without the executable extension).

This commit fixes these problems by establishing that the final
targets must indeed be the names of real files, and then keeping
track of the extensionless file names for other, meta uses.

Along the way, a few minor adjustments have also been made.