build: Remove the `clean' target from `libtommath/Makefile.in'
The file `libtommath/makefile.include', which is supplied by the
libtommath sources, already supplies its own `clean' target; besides
the fact that GNU `make' frowns upon multiple recipes for the same
target, it is likely the case that libtommath's own `clean' target
is worth using.

Indeed, the version of `clean' that appears in `Makefile.in' is
clearly a modified version of what appears in `makefile.include';
while the alterations are probably an improvement, they aren't so
worthwhile as to justify multiple definitions. One discrepancy of
note, though, is that the `Makefile.in' version lists `*.pdf' as
files to be deleted, while the other one does not, yet this doesn't
seem to make a difference.