build: libtommath/makefile.include: `arch' -> `uname -m'
Whenever building `libtommath', `make' would complain to me that
it could not find the `arch' command.

According to `info arch':

  `arch' prints the machine hardware name, and is equivalent to
  `uname -m'... `arch' is not installed by default, so portable
   scripts should not rely on its existence.

Well, this commit accepts that advice, replacing the invocation
of `arch' with the invocation of `uname -m'.

Now, it's true that this is a fix for what is essentially
another project, and this fix might one day make it "upstream",
but there's no reason that users of Dropbear must suffer until
then.