build: Improve support for out-of-tree builds
There's a lot going on here that makes it difficult to support
out-of-tree builds entirely, but this commit goes a long way:

    This commit allows for the out-of-tree building
    of the binaries of Dropbear and of the built-in
    `libtom{crypto,math}' dependencies.

That's basically all users of Dropbear care about.

Overall, the key is to ensure that `make' is given good values
for the `VPATH' and `srcdir' variables, and then also to use
`$(srcdir)' where appropriate.

Setting `VPATH' and `scrdir' is as easy as making sure that
the `configure' script is told about the various makefiles,
and including places in those makefiles to be so configured.

The irritating part is making sure that `$(srcdir)' is used
appropriately; the only place where this is really a problem is
in the makefiles for `libtommath', which was not written in a way
that supports out-of-tree builds in general (that is to say, while
it does work to build the binaries, it doesn't work to build the
ancillary files, such as the documentation).

Configure needs to know about the following files (which, at the
moment, simply provide for a working `clean' target):

  libtommath/etc/makefile
  libtommath/pics/makefile

This is achieved by adding them to `configure.ac' as files that
need to be configured; unfortunately, this seems to require
each of those files to end with the extension `.in'.

To keep the naming consistent, this commit renames each of those
files to `Makefile.in'.

At this point, neither of those files has been edited to include
a place for `configure' to insert a value for `VPATH' or `srcdir';
there are 2 reasons for this:

  * `git' will better be able to determine that these files have
    been renamed if they have not been modified at the same time.

  * The `clean' targets in them work without modification, and
    the rest of their targets are not useful to the main purpose
    of this commit: To make it possible to perform out-of-tree
    building of the *binaries*.

So, why not just copy those files into the out-of-tree build
directory without modifying those names or telling `configure'
about them? Well, because it's probably useful to prepare the
way for a possible future commit that does get the rest of the
targets working; then it will just be a matter of modifying
those files to do the right thing.