options: Explicitly set each boolean option to `1' or `0'
The C preprocessor's macro definition and expansion is essentially
just a matter of textual substitution. The consequence of this has
been that when one runs `make show-options', there appears output
like the following, which is helpful but a little too indirect for
good taste:

  DROPBEAR_SHA512 (((1)) || ((((1) || (1)))) || (((((1) || (1))))) || (0))

Specifically, this occurs because `DROPBEAR_SHA512' is defined in
the file `sysoptions.h' as follows:

  #define DROPBEAR_SHA512 ((DROPBEAR_SHA2_512_HMAC) || (DROPBEAR_ECC_521) \
                          || (DROPBEAR_SHA384) || (DROPBEAR_DH_GROUP16))

The values for each of the options on which `DROPBEAR_SHA512' depends
are simply substituded textually, yielding that large, unwieldy string
of ones, zeros, parentheses, and boolean operators.

That whole expression is only evaluated numerically by the C preprocessor
when it's used in an `#if' directive:

  $ git grep '#if DROPBEAR_SHA512'
  crypto_desc.c:#if DROPBEAR_SHA512
  libtomcrypt/src/headers/tomcrypt_custom.h:#if DROPBEAR_SHA512

One way to be more explicit about the value of an option such as
`DROPBEAR_SHA512' is to set its boolean value directly; rather
than define its value as a textual boolean expression, that
expression can instead be used as part of an `#if' directive
that ultimately determines whether the value to assign should be
`1' or `0'. In this case, the definition becomes:

  #if (DROPBEAR_SHA2_512_HMAC) || (DROPBEAR_ECC_521) \
      || (DROPBEAR_SHA384) || (DROPBEAR_DH_GROUP16)
         #define DROPBEAR_SHA512 1
  #else
         #define DROPBEAR_SHA512 0
  #endif

This transformation solves the problem, but it's a tad too verbose;
maybe that verbosity is worth it, though. Potential alternatives
include:

  * Keeping the old definitions, but replacing the explicit boolean
    operators with a system of fancy function-like macros.

  * Trying to evaluate such expressions as part of the recipe for
    `make show-options'.

However, each of those alternatives seems to add a certain complexity
that just isn't worth the trouble.