build: Wrap `config.h' in an include guard
If it's not useful for a header to be included multiple times, then
it seems like it should probably have an include guard.

In the case of `config.h', this is achieved by having `configure.ac'
create not `config.h', but rather `config-raw.h', which is then
included by a hand-constructed `config.h' with an include guard.