build: Correct the prerequisites
This commit provides 2 useful results:

  * It allows parallel make (`make -j') to function correctly.

  * It ensures that nearly every target (including every intermediate
    target) need be made only if at least one of that target's
    particular prerequisites has been touched.

    In other words, if you edit some header file, H, only those
    object files that include H will be re-made.

Before this commit, a number of header files, `$(HEADERS)', were listed
explicitly in Dropbear's makefile as prerequisites for the final,
complete program targets; however, the targets that *actually* depend on
such headers are not the complete programs, but rather the intermediate
object files from which those program targets are themselves made.

Thus, `$(HEADERS)' should actually be listed as the prerequisites for
the various intermediate object files. However, not every such object
file actually depends on every header, meaning that some object files
would be built unnecessarily if one of the headers is touched.

Therefore, `HEADERS' has been removed entirely, and replaced with
a mechanism by which each object file's prerequisites are determined
dynamically during the build process; the initial invocation of `make'
is only given a limited knowledge of the prerequisites, but every future
invocation has automatic access to fine[r]-grained details.

This is made possible with a command-line argument, which instructs the C
compiler or C preprocessor to output makefile-compatible rules that define
each object file's prerequisites. Fortunately, `llvm-gcc' appears to support
the most advanced such options, and even Tru64's C preprocessor supports the
most primitive option (according to the man pages that I have found).

When `configure' is run, the system is tested for the availability of these
options, and the best ones are chosen. If the system does not provide this
feature, then Dropbear's `Makefile.in' is set up to use a "static" set of
prerequisite rules, which was constructed in advance on a system that does
have the requisite capabilities.

Therefore, everyone should be covered.

To construct this static set of rules on a system which does support the
generation of dynamic prerequisites, one need only run the following command
in an out-of-tree build directory:

  $ make static-prereqs

Specifically:

  $ srcdir=/path/to/dropbear/source
  $ blddir=/path/to/dropbear/build # outside the source tree.
  $ cd "$srcdir"
  $ autoconf && autoheader
  $ mkdir -p "$blddir" && cd "$blddir"
  $ "$srcdir"/configure
  $ make static-prereqs

This places the new set of static rules in a single file which can be copied
to the same relative path under the Dropbear source tree:

  $ static_prereqs=build-system/prereqs/static/include.mk
  $ cp "$blddir/$static_prereqs" "$srcdir/$static_prereqs"

Of course, there is a subtle problem with this solution: It depends on the
C preprocessor being able to see all of the possible dependencies between
files, which can only occur if every possible `include' directive is visible
to the preprocessor; potentially, this might require that some options be
enabled, even though they wouldn't be by default, and yet enabling *all*
of the options with wild abandon might one day result in conflicts that
cause the preprocessor to choke.

Fortunately, experimentation has revealed that the `include' directives
associated with those options that are disabled by default seem to be
laid out such that they are visible, anyway; at the moment, there doesn't
seem to be a need to worry about tinkering with the options in order to
produce the most complete set of static prerequisites, but this will
require vigilance to keep that way as new features are added, or as the
source code is re-organized, etc; it would probably be wortwhile to
implement a more explicit means by which to reveal the connections.

The minimal knowledge of prerequisites to which `make' has access during the
initial run comprises the options header files, two of which need to be made
before almost everything else:

  localoptions.h
  default_options.h

Dropbear has only one object file that doesn't ultimately depend on those
headers:

  curve25519-donna.o

However, there's little worth in explaining this fact to `make'; instead,
the most practical solution is just to list these headers as prerequisites
for every object file.

It should be noted that the header file `localoptions.h' previously did
not need to exist; the makefiles would figure out whether it existed and
then set things up accordingly. While it's possible to keep this kind
of behavior, doing so would make it impossible to trigger the targets
of the `libtomcrypt' project when `localoptions.h' is *deleted*, unless
the same "runtime" discovery of prerequisites were also performed for
that builtin dependency. It's not worth the trouble; it's much easier
just to create an empty `localoptions.h' in the build directory so that
its presence can be expected and monitored.

Though such "runtime" discovery of prerequisites could be used for both
of the built-in libtom{crypt,math} dependencies, it turns out that these
projects are organized in a way that makes the results rather useless;
in particular, each `*.c' file includes all headers, so everything would
be re-made whenever any header is touched, which is what already happens
with the naive, manual approach that is currently employed.

Nevertheless, experimenting with this possibility has proven fruitful,
because it revealed a few dependencies that these projects share among
themselves, not only with respect to each other, but also with respect
to Dropbear's sources; for example, when Dropbear's `dbhelper.h' is
touched, both `libtomcrypt' and `libtommath' have object files that
need to be re-made, etc.

The responsiblity for managing such prerequisites has been moved
from Dropbear's makefile to the makefiles of the libtom* libraries;
because Dropbear's makefile merely runs a new instance of `make'
to construct each one of those libraries, only the makefiles of
those libraries can handle the prerequisites meaningfully.

Nevertheless, there is one prerequisite of libtomcrypt that must
*also* be handled by Dropbear's makefile: `default_options.h'. This
is because allowing libtomcrypt's makefile to drive the creation
of that file would break one of the assumptions of `make', namely
that a particular target is only ever made by running one recipe for
it at a time; during a parallel make, this interplay between the
makefiles would escape the notice of each separate instance of `make',
thereby creating a situation where two of them are trying to make
`default_options.h' at the same time.

The same problem also exists for `localoptions.h', but only as a
tehcnicality; currently, there is actually no risk of a practical
data race, because the result of any such race is still just an
empty file whose timestamp indicates that targets should be
re-made.

To keep things simple, the makefiles have been set up such that only
Dropbear's makefile is allowed to orchestrate `make'; the user is not
allowed to run libtomcrypt's makefile directly, and the user is
warned loudly if the user tries to do so.