build: Upgrade and use `install-sh'
According to `info autoconf', it's valuable to include and use a
copy of their `install-sh' script as the `$(INSTALL)' program.

The Dropbear source directory has had a copy of an older version
of this script, but it hasn't been used. This commit upgrades
that script to the one provided by `AutoMake 1.5', and then
alters `configure.ac' to use it.