options: Remove public traces of `SCPPROGRESS'
The actual "option" in question is `PROGRESS_METER'; defining the
`SCPPROGRESS=1' make variable on the command line was, has been, and
remains a way (albeit an obsolete way) to define `PROGRESS_METER'.

Now, `PROGRESS_METER' is officially documented in `default_options.h.in',
and it should be defined accordingly.

Also, the conditional directive in `progressmeter.c' has been moved down
below the include directives; not only is this consistent with the way
other files are laid out, but it also exposes the relevant header files
to programmatic inspection by the C Preprocessor, which will be useful
for a future a commit that enables dynamically generated makefile
prerequisites.